+++ src/Command.php
@@ -5,12 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Oracle;
 
 use PDO;
-use PDOException;
-use Throwable;
 use Yiisoft\Db\Driver\Pdo\AbstractPdoCommand;
-use Yiisoft\Db\Exception\ConvertException;
 use Yiisoft\Db\QueryBuilder\AbstractQueryBuilder;
-use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
 
 use function array_keys;
@@ -91,11 +87,6 @@ final class Command extends AbstractPdoCommand
         return $this->setSql($sql)->queryColumn();
     }
 
-    protected function getQueryBuilder(): QueryBuilderInterface
-    {
-        return $this->db->getQueryBuilder();
-    }
-
     protected function bindPendingParams(): void
     {
         $paramsPassedByReference = [];
@@ -114,41 +105,6 @@ final class Command extends AbstractPdoCommand
                 );
             } else {
                 $this->pdoStatement?->bindValue($name, $value->getValue(), $value->getType());
-            }
-        }
-    }
-
-    /**
-     * @psalm-suppress UnusedClosureParam
-     *
-     * @throws Throwable
-     */
-    protected function internalExecute(?string $rawSql): void
-    {
-        $attempt = 0;
-
-        while (true) {
-            try {
-                if (
-                    ++$attempt === 1
-                    && $this->isolationLevel !== null
-                    && $this->db->getTransaction() === null
-                ) {
-                    $this->db->transaction(
-                        fn () => $this->internalExecute($rawSql),
-                        $this->isolationLevel
-                    );
-                } else {
-                    $this->pdoStatement?->execute();
-                }
-                break;
-            } catch (PDOException $e) {
-                $rawSql = $rawSql ?: $this->getRawSql();
-                $e = (new ConvertException($e, $rawSql))->run();
-
-                if ($this->retryHandler === null || !($this->retryHandler)($e, $attempt)) {
-                    throw $e;
-                }
             }
         }
     }