+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bug #233: Refactor `DMLQueryBuilder`, related with yiisoft/db#746 (@Tigrov)
 - Enh #230: Improve column type #230 (@Tigrov)
 - Bug #240: Remove `RECURSIVE` expression from CTE queries (@Tigrov)
+- Enh #243: Move methods from `Command` to `AbstractPdoCommand` class (@Tigrov)
 
 ## 1.1.0 July 24, 2023